refactor(i18n): clarify main process message catalog
