+++ src/main/locale/main-process-messages.ts
